fix printf with 64bit time_t Printing long long with %ld is UB. Converting long long to long could overflow. To be compatible with 32bit time_t we need an explicit cast to long long.

+++ tests/ts_print_mt.c
@@ -241,10 +241,10 @@ int main(int argc, char **argv)
 				if (!(samp_mt[j][i].valid & TSLIB_MT_VALID))
 					continue;
 
-				printf(YELLOW "sample %d - %ld.%06ld -" RESET " (slot %d) %6d %6d %6d\n",
+				printf(YELLOW "sample %d - %lld.%06lld -" RESET " (slot %d) %6d %6d %6d\n",
 				       j,
-				       samp_mt[j][i].tv.tv_sec,
-				       samp_mt[j][i].tv.tv_usec,
+				       (long long)samp_mt[j][i].tv.tv_sec,
+				       (long long)samp_mt[j][i].tv.tv_usec,
 				       samp_mt[j][i].slot,
 				       samp_mt[j][i].x,
 				       samp_mt[j][i].y,
